Fix: Url을 읽어오는 작업은 어떠한 메서드든 공통이기에 interface static 메서드로 이동
