+++ src/test/java/reader/RequestReaderTest.java
@@ -1,11 +1,14 @@
 package reader;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import request.HttpRequest;
 import util.HttpMethod;
 import util.error.HttpsErrorMessage;
 
 import java.net.ProtocolException;
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -26,5 +29,13 @@ class RequestReaderTest {
     void selectRequestReaderByMethodError() throws ProtocolException {
         ProtocolException protocolException = assertThrows(ProtocolException.class, () -> RequestReader.selectRequestReaderByMethod(HttpMethod.NOT_MATCH));
         assertThat(protocolException.getMessage()).isEqualTo(HttpsErrorMessage.NOT_VALID_HTTP_FORMAT);
+    }
+
+    @Test
+    @DisplayName("Get 메서드의 요청의 경우 url파싱 테스트")
+    void findPathInRequest() {
+        final String requestHeader = "GET /index.html HTTP/1.1";
+        HttpRequest httpRequest = new HttpRequest(List.of(requestHeader));
+        Assertions.assertThat(RequestReader.findPathInRequest(httpRequest)).isEqualTo("/index.html");
     }
 }
